Move HasDynamicRelationships trait under Modules\Core\Traits namespace
